# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "電商平台 Auth API",
        "description": "認證相關 API 文件範例",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chat/users": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "取得其他會員列表",
                "description": "取得其他會員列表",
                "operationId": "b1e19c4f53ac698318a47355befd2150",
                "responses": {
                    "200": {
                        "description": "成功取得會員列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/room": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "取得或建立聊天室",
                "description": "取得或建立私聊聊天室",
                "operationId": "d56f7dfeeb545a4505d10be13c9e95ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功取得或建立聊天室",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "room_id": {
                                            "type": "integer"
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/rooms/{roomId}/messages": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "取得聊天室歷史訊息",
                "description": "取得聊天室歷史訊息",
                "operationId": "f25b058a5a5e7692f487bd3f87b2a76c",
                "parameters": [
                    {
                        "name": "roomId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功取得訊息",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "message": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/messages": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "發送訊息",
                "description": "發送訊息",
                "operationId": "3abf37fffdcd7415ddb9c11a9e243a11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "message"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功發送訊息",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "room_id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/rooms/{roomId}/read": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "標記已讀",
                "description": "標記已讀",
                "operationId": "01d45437fbd78edf37125eaf371edd50",
                "parameters": [
                    {
                        "name": "roomId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功標記已讀"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/unread-counts": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "取得未讀數",
                "description": "取得未讀數",
                "operationId": "9601dbcb059c4b635741f8624180c453",
                "responses": {
                    "200": {
                        "description": "成功取得未讀數",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "room_id": {
                                            "type": "integer"
                                        },
                                        "count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}